Move analysis command into their own subspace
