+++ freqtrade/optimize/analysis/lookahead_analysis.py
