+++ freqtrade/optimize/analysis/recursive_analysis.py
